renamed hello servlet and changed mesage, mounted it on root
